🐛  (next) Image.client, and optimizations [b]
What a great idea to turn optimizations off, haha.
This should do the thing for LCP/FCP, and then
 we can continue to tweak the `sizes` as appropriate
