+++ sites/jeromefitzgerald.com/src/components/Notion/Blocks/Embed.Twitter.tsx
@@ -1,4 +1,4 @@
-import 'server-only'
+'use client'
 
 import { Caption } from '@jeromefitz/ds/components/Caption'
 import { NotionEmoji as EmojiWrapper } from 'next-notion/src/blocks/Emoji'
